fix: Use a unique ID for each workflow to avoid S3 stale entries
Github allows us use a unique worflow id. Until now, if we attempted
to build a firmware with the same repo/type/version (because we may
added some changes in our app repo) the old S3 files from previous
workflows would be included to the next ones. For example, if one we-
re running the same workflow, first for esp32s3 and then for esp32s2,
the esp32s3 binary files would still be there, although they would be
deprecated. We solve that by introducing the unique github workflow
id to S3 namespacing.

Signed-off-by: Ilias Lagomatis <ilago@nubificus.co.uk>